Controller authorization

+++ src/main/java/com/hodolog/api/controller/PostController.java
@@ -56,20 +56,18 @@ public class PostController {
 
     private final PostService postService;
     // 유연한 대응 필요
-    @PostMapping("/posts")
-    public Map<String, String> post(@RequestBody @Valid PostCreate request/*, BindingResult result*/) {
-        // 데이터를 검증하는 이유
-        // 1. 매번 메서드마다 값을 검증 해야한다.(반복)
-        // 검증 부분에서 버그가 발생할 여지가 높음.
-        // 2. 응답값에 HashMap -> 응답 클래스를 만들어주는게 좋다.
-        // 3. 여러개의 에러처리 힘듬
-        // 4. 세 번이상의 반복 작업은 피해야한다.
-            //자동화 고려
-        // 1. client 개발자가 깜빡할 수 있다. 실수로 값을 안보낼수 있다
-        // 2. client bug로 값이 누락될 수 있다.
-        // 3. 외부에 해커가 값을 임의로 조작해서 보낼 수 있다.
-        // 4. DB에 값을 저장할 때 의도치 않은 오류가 발생할 수 있다.
-        // 5. 서버 개발자의 편안함을 위해서
+    // 데이터를 검증하는 이유
+    // 1. 매번 메서드마다 값을 검증 해야한다.(반복)
+    // 검증 부분에서 버그가 발생할 여지가 높음.
+    // 2. 응답값에 HashMap -> 응답 클래스를 만들어주는게 좋다.
+    // 3. 여러개의 에러처리 힘듬
+    // 4. 세 번이상의 반복 작업은 피해야한다.
+    //자동화 고려
+    // 1. client 개발자가 깜빡할 수 있다. 실수로 값을 안보낼수 있다
+    // 2. client bug로 값이 누락될 수 있다.
+    // 3. 외부에 해커가 값을 임의로 조작해서 보낼 수 있다.
+    // 4. DB에 값을 저장할 때 의도치 않은 오류가 발생할 수 있다.
+    // 5. 서버 개발자의 편안함을 위해서
 //        if(result.hasErrors()){
 //            List<FieldError> fieldErrors = result.getFieldErrors();
 //            FieldError firstFieldError = fieldErrors.get(0);
@@ -82,12 +80,10 @@ public class PostController {
 //
 //        }
 //        log.info("params={}",params.toString());
-        // repository.save(params)
-        //postService.write(request) -> Controller -> Service -> Repository -> 최종적으로 넘어온 Json 데이터 값을 Post Entity 형태로 변환하여 저장;
-        request.validate();
-        postService.write(request);
-        return Map.of();
-//        String title = params.getTitle();
+    // repository.save(params)
+    //postService.write(request) -> Controller -> Service -> Repository -> 최종적으로 넘어온 Json 데이터 값을 Post Entity 형태로 변환하여 저장;
+
+    //        String title = params.getTitle();
 //        if(title == null || title.equals("")){
 //            //error
 //            // 1. 빡세다. (노가다)
@@ -104,10 +100,24 @@ public class PostController {
 //            //error
 //        }
 
-        /*
-        posts -> 글 전체 조회(검색 + 페이징)
-                posts/{postId} -> 글 한개만 조회
-        */
+    /*
+    posts -> 글 전체 조회(검색 + 페이징)
+            posts/{postId} -> 글 한개만 조회
+    */
+    @PostMapping("/posts")
+    public Map<String, String> post(@RequestBody @Valid PostCreate request, @RequestHeader String authorization) {
+
+        //기본적인 요청 인증 값 확인
+        //1. GET Parameter -> ??
+        if (authorization.equals("hodolman")) {
+            request.validate();
+            postService.write(request);
+        }
+        //2. Header
+
+//        request.validate();
+//        postService.write(request);
+        return Map.of();
 
     }
 
@@ -131,13 +141,19 @@ public class PostController {
     }
 
     @PatchMapping("/posts/{postId}")
-    public void edit(@PathVariable Long postId, @RequestBody @Valid PostEdit request)  {
-        postService.edit(postId, request);
+    public void edit(@PathVariable Long postId, @RequestBody @Valid PostEdit request, @RequestHeader String authorization)  {
+        if (authorization.equals("hodolman")) {
+            postService.edit(postId, request);
+        }
+
     }
 
     @DeleteMapping("/posts/{postId}")
-    public void delete(@PathVariable Long postId)  {
-        postService.delete(postId);
+    public void delete(@PathVariable Long postId, @RequestHeader String authorization)  {
+        if (authorization.equals("hodolman")) {
+            postService.delete(postId);
+        }
+
     }
 
 }
